prepare multiple controllers

+++ bitser/src/main/java/com/github/knokko/bitser/connection/ClientStream.java
@@ -13,7 +13,7 @@ public interface ClientStream {
 
 	interface Factory {
 
-		ClientStream createStream(int controllerID);
+		ClientStream createStream(long controllerID);
 	}
 
 	interface InputReader {
